+++ src/application.tsx
@@ -11,6 +11,7 @@ import { routes } from './config/routes'
 import { support_routes } from './config/support-routes'
 import { user_dashboard_routes } from './config/user-dashboard-routes'
 import { useScrollToTop } from './hooks/use-scroll-to-top'
+import { useTokenRefresh } from './hooks/use-token-refresh'
 import CreatorGuideOverlay from './components/shared/creator-guide-overlay'
 import CreatorDashboardLayout from './layouts/creator-dashboard-layout'
 import IndexLayout from './layouts/root-layout'
@@ -36,6 +37,7 @@ const queryClient = new QueryClient({
 
 function AppRoutes() {
   useScrollToTop()
+  useTokenRefresh()
 
   return (
     <>